% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterComrades.R
\docType{methods}
\name{clusterComrades}
\alias{clusterComrades}
\alias{clusterComrades,comradesDataSet-method}
\title{clusterComrades}
\usage{
clusterComrades(cds, cores = 3, stepCount = 2, clusterCutoff = 20)
}
\arguments{
\item{cds}{comradesDataSet object created with comradesDataSet}

\item{cores}{numeric - The number of cores to use}

\item{stepCount}{Stringency for clustering}

\item{clusterCutoff}{The minimum number of reads a cluster requires}
}
\description{
This method clusters the duplexes.
}
