% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterGrangesList<-}
\alias{clusterGrangesList<-}
\alias{clusterGrangesList<-,rnaCrosslinkDataSet-method}
\title{clusterGrangesList<-}
\usage{
clusterGrangesList(x) <- value
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}

\item{value}{A replacement}
}
\value{
No return - Sets a new clusterGrangesList slot
}
\description{
Set new clusterGrangesList slot
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

newclusterGrangesList <- clusterGrangesList(cds)
clusterGrangesList(cds) <- newclusterGrangesList
}
