% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterTableList}
\alias{clusterTableList}
\alias{clusterTableList,comradesDataSet-method}
\title{clusterTableList}
\usage{
clusterTableList(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the cluster coordinates in data frame format
}
