% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterTableList<-}
\alias{clusterTableList<-}
\alias{clusterTableList<-,comradesDataSet-method}
\title{clusterTableList<-}
\usage{
clusterTableList(x) <- value
}
\arguments{
\item{x}{A comradesDataSet object}

\item{value}{A replacement}
}
\description{
Set new clusterTableList slot
}
