% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterComradesMethodsAndHelpers.R
\docType{methods}
\name{compareKnown}
\alias{compareKnown}
\alias{compareKnown,comradesDataSet-method}
\title{compareKnown}
\usage{
compareKnown(trimmedClusters, knownMat, type)
}
\arguments{
\item{trimmedClusters}{a \code{comradesDataSet} object, 
run \code{trimClusters()} on the  \code{comradesDataSet} first}

\item{knownMat}{Matrix - A marix(ncol = lengthRNA,nrow = lengthRNA) where a
value in matrix[x,y] would indicate a known interation between nucleotide 
x and nucleotide y}

\item{type}{string - the Analysis stage of clusters you would like to compare you can find 
available types by just running the objects name}

\item{rna}{The ID of the RNA of interest (can use rna(cdsObject))}
}
\value{
Returns a \code{comradesClusteredDataSet} object

The 3 attributes matrixList, clusterTableList and clusterGrangesList 
will gain the \code{types} "known" and "novel" and "knownAndNovel"
}
\description{
This method compares the current object to a know structure.run 
\code{trimClusters()} on the  \code{comradesDataSet} first
}
