% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldComradesMethodsAndHelpers.R
\docType{methods}
\name{compareKnownStructures}
\alias{compareKnownStructures}
\alias{compareKnownStructures,comradesDataSet-method}
\title{compareKnownStructures}
\usage{
compareKnownStructures(foldedCds, file)
}
\arguments{
\item{foldedCds}{\code{comradesDataSet} after running foldComrades}

\item{file}{a two column file with column header i and j with numeric values showing
nucleoide i binds to nucleotide j}
}
\value{
Returns a dataframe
}
\description{
this method prints some feature level statistics
}
