% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesOO.R, R/comradesOO-package.R
\docType{package}
\name{comradesOO}
\alias{comradesOO}
\alias{comradesOO-package}
\title{comradesOO: A package for analysing a COMRADES dataset}
\description{
The package consists of 1 Object:
   comradesDataSet

The package consists of 1 Object:
   comradesDataSet
}
\section{\code{\link{comradesDataSet}}}{
NA

NA
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JLP-BioInf/comradesOO/issues}
}


Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JLP-BioInf/comradesOO/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Price \email{jlp76@cam.ac.uk}

}
\keyword{internal}
