% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{featureInfo}
\alias{featureInfo}
\alias{featureInfo,comradesDataSet-method}
\title{featureInfo}
\usage{
featureInfo(cds)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}
}
\value{
A list - Feature level and transcript level counts for each sample
}
\description{
Produces a list list of 2 elemnts 'transcript' and 'family'
Each element contains a table with the counts for each RNA in each sample.
}
\examples{
cds = makeExampleComradesDataSet()
sampleTable(cds)

}
