% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{findBasePairsRNAcoFold2}
\alias{findBasePairsRNAcoFold2}
\title{findBasePairsRNAcoFold2}
\usage{
findBasePairsRNAcoFold2(
  startPos1,
  endPos1,
  seq1,
  startPos2,
  endPos2,
  seq2,
  fasta,
  shape
)
}
\arguments{
\item{startPos1}{Start of the cluster side x}

\item{endPos1}{End of the cluster side x}

\item{seq1}{Sequence of x}

\item{startPos2}{Start of the cluster side y}

\item{endPos2}{End of the cluster side y}

\item{seq2}{Sequence of y}

\item{fasta}{\code{rnaRefs}}

\item{shape}{shape reactivities}
}
\value{
A table of clusters and coordinates with folds
}
\description{
Folds the clusters using Vienna RNAfold
}
