% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{findBasePairsRNAfold}
\alias{findBasePairsRNAfold}
\title{findBasePairsRNAfold}
\usage{
findBasePairsRNAfold(startPos, endPos, seqs, fasta, shape)
}
\arguments{
\item{startPos}{Start of the cluster side x}

\item{endPos}{End of the cluster side x}

\item{seqs}{Sequence of x}

\item{fasta}{\code{rnaRefs}}

\item{shape}{shape reactivities}
}
\value{
A table of clusters and coordinates with folds
}
\description{
Folds the clusters using Vienna RNA duplex
}
