% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldComrades.R
\docType{methods}
\name{foldComrades}
\alias{foldComrades}
\alias{foldComrades,comradesDataSet-method}
\title{foldComrades}
\usage{
foldComrades(
  cdsObject,
  rnaRefs,
  start,
  end,
  evCutoff = 1,
  ensembl = 50,
  constraintNumber = 20,
  shape = 0
)
}
\arguments{
\item{cdsObject}{comradesDataSet object created with comradesDataSet}

\item{rnaRefs}{named List - a list with named elements that correspond to the
.RNA of interest. The element of the list must be a fasta file that has
been read with \code{seqinr::read.fasta()}}

\item{start}{Start of segmnent to fold}

\item{end}{End of segmnent to fold}

\item{evCutoff}{Mininum number of read support for contraint to be included in folding}

\item{ensembl}{Number of structures to Nake}

\item{constraintNumber}{Number of constraints to add to each final fold}

\item{shape}{shape reactivities (0 for no constraints)}
}
\description{
This methods folds an ensebl of structures for the whole RNA or chosen region
of the RNA. See \code{comradesDataSet} for slot information.
}
