% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonHelpersAndMethods.R
\docType{methods}
\name{getAdjacancyMat}
\alias{getAdjacancyMat}
\title{getAdjacancyMat}
\usage{
getAdjacancyMat(InputGranges, nucletideOrPerc, cutoff)
}
\arguments{
\item{InputGranges}{list created with InputToGRanges (but just the gap section of the list)}

\item{nucletideOrPerc}{measure difference by percentage or nucleotides}

\item{cutoff}{The maximum difference before giving these two gaps 0}
}
\value{
A list of Adjacancy matrices
}
\description{
Makes and adjacency matrix list (for clustering)
}
\details{
Makes and adjacency matrix list (for clustering)
}
