% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{getClusterClusterShortRangeWhole}
\alias{getClusterClusterShortRangeWhole}
\title{getClusterClusterShortRangeWhole}
\usage{
getClusterClusterShortRangeWhole(cluster, seq)
}
\arguments{
\item{cluster}{cluster positions}

\item{seq}{sequence of transcript}
}
\value{
The same table with an extra column
}
\description{
Decides if a cluster is long or short range
then either grabs the whole sequence or the sequence of the two sides of the
interaction separately.
}
