% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{getInteractions}
\alias{getInteractions}
\alias{getInteractions,comradesDataSet-method}
\title{getInteractions}
\usage{
getInteractions(cds, interactor)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{interactor}{The rna to show interactions with}
}
\value{
A table showign the read coverage of the interacting RNA
}
\description{
This method returns a table interactions of an RNA (interactor) on the RNA of
interest use topInteracters.
}
\examples{
cds = makeExampleComradesDataSet()
getInteractions(cds, 'transcript2')

}
