% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{getReverseInteractions}
\alias{getReverseInteractions}
\alias{getReverseInteractions,comradesDataSet-method}
\title{getReverseInteractions}
\usage{
getReverseInteractions(cds, interactor)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{interactor}{The rna to show interactions with}
}
\description{
This method prints interactions of the RNA of interest on another RNA
transcript.
}
