% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{group}
\alias{group}
\alias{group,comradesDataSet-method}
\title{group}
\usage{
group(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the indeces for each group for the instance
}
