% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{hybFiles}
\alias{hybFiles}
\alias{hybFiles,comradesDataSet-method}
\title{hybFiles}
\usage{
hybFiles(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the data in original format
}
