% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterComradesMethodsAndHelpers.R
\docType{methods}
\name{hybToGRanges}
\alias{hybToGRanges}
\title{hybToGRanges}
\usage{
hybToGRanges(hybList, rna)
}
\arguments{
\item{hybList}{the original hybList created with readHybFiles or subsetHybList}

\item{rna}{The rna of interest}
}
\value{
A list of GRanges data in hyb format
}
\description{
This function is useful to turn a list of hyb data into lists of GRanges
It creates a list for each sample one for the left side one for the right
side and one for the gap in the middle.
}
