% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{InputFiles}
\alias{InputFiles}
\alias{InputFiles,rnaCrosslinkDataSet-method}
\title{InputFiles}
\usage{
InputFiles(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A list of tables in the original input format, one entry for each sample
}
\description{
Extract the data in original format
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

InputFiles(cds)
}
