% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{InputToGRanges}
\alias{InputToGRanges}
\title{InputToGRanges}
\usage{
InputToGRanges(InputList, rna)
}
\arguments{
\item{InputList}{the original InputList created with readInputFiles or subsetInputList}

\item{rna}{The rna of interest}
}
\value{
A list of GRanges data in Input format
}
\description{
This function is useful to turn a list of Input data into lists of GRanges
It creates a list for each sample one for the left side one for the right
side and one for the gap in the middle.
}
