% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\name{makeExampleComradesDataSet}
\alias{makeExampleComradesDataSet}
\title{makeExampleComradesDataSet}
\usage{
makeExampleComradesDataSet()
}
\value{
An example comradesDataSet objext
}
\description{
Creat a minimal example comradesdataSetObject
}
