% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\name{makeExamplernaCrosslinkDataSet}
\alias{makeExamplernaCrosslinkDataSet}
\title{makeExamplernaCrosslinkDataSet}
\usage{
makeExamplernaCrosslinkDataSet()
}
\value{
An example rnaCrosslinkDataSet objext
}
\description{
Creat a minimal example rnaCrosslinkdataSetObject
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
}
