% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{matrixList}
\alias{matrixList}
\alias{matrixList,comradesDataSet-method}
\title{hybFiles}
\usage{
matrixList(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the contact matrices
}
