% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{matrixList<-}
\alias{matrixList<-}
\alias{matrixList<-,comradesDataSet-method}
\title{matrixList}
\usage{
matrixList(x) <- value
}
\arguments{
\item{x}{A comradesDataSet object}

\item{value}{A replacement}
}
\description{
Set new matrixList slot
}
