% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldComradesMethodsAndHelpers.R
\docType{methods}
\name{plotComparisonArc}
\alias{plotComparisonArc}
\alias{plotComparisonArc,comradesDataSet-method}
\title{plotComparisonArc}
\usage{
plotComparisonArc(foldedCds, s1 = "s1", s2 = "s2", n1 = 1, n2 = 2)
}
\arguments{
\item{foldedCds}{\code{comradesDataSet} after running foldComrades}

\item{s1}{sample of structure 1}

\item{s2}{sample of structure 2}

\item{n1}{number of structure from first sample}

\item{n2}{number of structure from first sample}
}
\description{
This method plots two structures chosen from the
plotEnsemblePCA method
}
