% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldComradesMethodsAndHelpers.R
\docType{methods}
\name{plotEnsemblePCA}
\alias{plotEnsemblePCA}
\alias{plotEnsemblePCA,comradesDataSet-method}
\title{plotEnsemblePCA}
\usage{
plotEnsemblePCA(foldedCds, labels = F, split = T)
}
\arguments{
\item{foldedCds}{\code{comradesDataSet} after running foldComrades}

\item{labels}{plot with labels or not (TRUE/FALSE)}

\item{split}{split the plot using facets based on the samples  (TRUE/FALSE)}
}
\description{
This method plots a PCA of the ensembl
}
