% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{printClustersFast}
\alias{printClustersFast}
\title{printClustersFast}
\usage{
printClustersFast(dir, clustering, highest_clusters, left, right)
}
\arguments{
\item{dir}{the directory that contains the *Inputrids.Input files}

\item{clustering}{The output from the iGraph function cluster_walktrap for the (made with adjacency matrix input)}

\item{highest_clusters}{The cluster you are interested in keeping}

\item{left}{list created with InputToGRanges (but just the left section of the list)}

\item{right}{list created with InputToGRanges (but just the right section of the list)}
}
\value{
A table of clusters and coordinates
}
\description{
Makes a table with the coordinates of the clusters
}
\details{
Does the same as printClusters but is a lot faster and does not create plots
of each cluster
}
