% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonStatsAndPlots.R
\docType{methods}
\name{readNumbers}
\alias{readNumbers}
\alias{readNumbers,comradesDataSet-method}
\title{readNumbers}
\usage{
readNumbers(knowClusteredCds, rna)
}
\arguments{
\item{knowClusteredCds}{A comradesDataSet object after clustering has been performed}

\item{rna}{The RNA ID of interest - use rna(cdsObject).}
}
\description{
This method prints a table showing the number of duplexes in the clusters in each step of the analysis
}
