% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkOO.R, R/rnaCrosslinkOO-package.R
\docType{package}
\name{rnaCrosslinkOO}
\alias{rnaCrosslinkOO}
\title{rnaCrosslinkOO: A package for analysing a rnaCrosslink dataset}
\description{
The package consists of 1 Object:
   rnaCrosslinkDataSet

The package consists of 1 Object:
   rnaCrosslinkDataSet
}
\section{\code{\link{rnaCrosslinkDataSet}}}{
NA

NA
}

\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JLP-BioInf/rnaCrosslinkOO/issues}
}


Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/JLP-BioInf/rnaCrosslinkOO/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Price \email{jlp76@cam.ac.uk} (\href{https://orcid.org/0000-0001-6554-5667}{ORCID})

Other contributors:
\itemize{
  \item Andrew Lim \email{ajhl3@cam.ac.uk} [contributor]
}

}
\keyword{internal}
