% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkQC.R
\docType{methods}
\name{rnaCrosslinkQC}
\alias{rnaCrosslinkQC}
\title{rnaCrosslinkQC}
\usage{
rnaCrosslinkQC(sampleTable, directory)
}
\arguments{
\item{sampleTable}{string - The address of the sample table, the sample table
must have 4 columns, fileName (the full path and file name of the input
Input file for each sample ), group ("s" - sample or "c" - control),
sample (1,2,3, etc), sampleName (must be unique).}

\item{Directory}{A directory address to write the files}
}
\value{
ggplot and txt file
}
\description{
get a plot fo the read lengths and transcripts in the dataset
The fucntion will make 1 pdf and 2 text file in the directory provided
}
