% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{rnas}
\alias{rnas}
\alias{rnas,comradesDataSet-method}
\title{rnas}
\usage{
rnas(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\value{
A character - the ID of the RNA
}
\description{
Extract the rna ID for the instance
}
