% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{rnaSize}
\alias{rnaSize}
\alias{rnaSize,comradesDataSet-method}
\title{rnaSize}
\usage{
rnaSize(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\value{
A numeric - the size of the RNA
}
\description{
Extract the size of the RNA for the instance
}
