% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterrnaCrosslinkMethodsAndHelpers.R
\name{sampleChimeras}
\alias{sampleChimeras}
\title{sampleChimeras}
\usage{
sampleChimeras(chimeraList)
}
\arguments{
\item{chimeraList}{list of chimeras}
}
\description{
This function samples chimeras into smaller chunks so that clustering is 
quicker
}
