% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{sampleTable}
\alias{sampleTable}
\alias{sampleTable,comradesDataSet-method}
\title{sampleTable}
\usage{
sampleTable(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the sample table for the instance
}
