% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterComradesMethodsAndHelpers.R
\docType{methods}
\name{subsetHybList2}
\alias{subsetHybList2}
\title{subsetHybList2}
\usage{
subsetHybList2(hybList, min, max, length)
}
\arguments{
\item{hybList}{the original hybList created with readHybFiles}

\item{min}{the rna of interest that you want to subset}

\item{max}{The number of randomly subsetted chimeric reads you need}

\item{length}{The number of randomly subsetted chimeric reads you need}
}
\value{
A list of subsetted hyb files
}
\description{
Subset a list of hyb files
}
\details{
Function used to subset a list of hyb data created by readHybFiles
This function produces the same size list as before but 
it returns ONLY the rna of interest and also
Choose duplexes where the nt difference in position between the
one side and other side of an interaction is between min and max
}
