% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterrnaCrosslinkMethodsAndHelpers.R
\docType{methods}
\name{subsetInputList2}
\alias{subsetInputList2}
\title{subsetInputList2}
\usage{
subsetInputList2(InputList, min, max, length)
}
\arguments{
\item{InputList}{the original InputList created with readInputFiles}

\item{min}{the rna of interest that you want to subset}

\item{max}{The number of randomly subsetted chimeric reads you need}

\item{length}{The number of randomly subsetted chimeric reads you need}
}
\value{
A list of subsetted Input files
}
\description{
Subset a list of Input files
}
\details{
Function used to subset a list of Input data created by readInputFiles
This function produces the same size list as before but 
it returns ONLY the rna of interest and also
Choose duplexes where the nt difference in position between the
one side and other side of an interaction is between min and max
}
