% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{swapHybs2}
\alias{swapHybs2}
\title{swapHybs2}
\usage{
swapHybs2(hybList, rna)
}
\arguments{
\item{hybList}{the original hybList created with readHybFiles or subsetHybList}

\item{rna}{The rna of interest}
}
\value{
A list of "swapped" hyb data
}
\description{
Swap the table to ensure that 3 prime most duplex side is on the left of the table
used to make one sides heatmaps and other reasons where having the left of the table
coming after the right side is a problem. Different from swapHybs as it
ensure that BOTH duplex sides originate from the RNA of interest.
}
