% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{swapInputs3}
\alias{swapInputs3}
\title{swapInputs3}
\usage{
swapInputs3(InputList, rna)
}
\arguments{
\item{InputList}{the original InputList created with readInputFiles or subsetInputList}

\item{rna}{The rna of interest}
}
\value{
A list of "swapped" Input datas
}
\description{
Swap the table to ensure that 3 prime most duplex side is ont he left of the table
used to make one sides heatmaps and other reasons where having the left of the table
coming after the right side is a problem. Different from swapInputs as it
ensure that BOTH duplex sides originate from the RNA of interest.
}
