% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{topTranscripts}
\alias{topTranscripts}
\alias{topTranscripts,comradesDataSet-method}
\title{topTranscripts}
\usage{
topTranscripts(cds, ntop = 10)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{ntop}{the number of entries to display}
}
\value{
A table, the number of counts per sample per transcript
}
\description{
This method prints the top transcripts that have the most duplexes
assigned
}
\examples{
cds = makeExampleComradesDataSet()
topTranscripts(cds)
}
