% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterGrangesList}
\alias{clusterGrangesList}
\alias{clusterGrangesList,comradesDataSet-method}
\title{clusterGrangesList}
\usage{
clusterGrangesList(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\description{
Extract the cluster coordinates in granges format
}
