% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterGrangesList}
\alias{clusterGrangesList}
\alias{clusterGrangesList,rnaCrosslinkDataSet-method}
\title{clusterGrangesList}
\usage{
clusterGrangesList(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A list of Granges objects showing the positions of each cluster, one entry for each sample
}
\description{
Extract the cluster coordinates in granges format
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

clusterGrangesList(cds)
}
