% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterTableList}
\alias{clusterTableList}
\alias{clusterTableList,rnaCrosslinkDataSet-method}
\title{clusterTableList}
\usage{
clusterTableList(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A list of tables showing the vienna structures of each cluster
}
\description{
Extract the cluster coordinates in data frame format
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

clusterTableList(cds)
}
