% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{clusterTableList<-}
\alias{clusterTableList<-}
\alias{clusterTableList<-,rnaCrosslinkDataSet-method}
\title{clusterTableList<-}
\usage{
clusterTableList(x) <- value
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}

\item{value}{A replacement}
}
\value{
No return - Sets a new clusterTableList slot
}
\description{
Set new clusterTableList slot
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

newclusterGrangesList <- clusterTableList(cds)
clusterTableList(cds) <- newclusterGrangesList
}
