% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{featureInfo}
\alias{featureInfo}
\alias{featureInfo,comradesDataSet-method}
\title{featureInfo}
\usage{
featureInfo(cds, file)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{file}{Directory for output files}
}
\description{
this method prints some feature level statistics
}
