% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{featureInfo}
\alias{featureInfo}
\alias{featureInfo,rnaCrosslinkDataSet-method}
\title{featureInfo}
\usage{
featureInfo(cds)
}
\arguments{
\item{cds}{a \code{rnaCrosslinkDataSet} object}
}
\value{
A list - Feature level and transcript level counts for each sample
}
\description{
Produces a list list of 2 elemnts 'transcript' and 'family'
Each element contains a table with the counts for each RNA in each sample.
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
featureInfo(cds)

}
