% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,rnaCrosslinkDataSet-method}
\title{getData}
\usage{
getData(x, data, type)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}

\item{data}{The data type to return <InputFiles | matrixList | clusterGrangesList | clusterTableList>}

\item{type}{The analysis stage <original | noHost | originalClusters | trimmedClusters>}
}
\value{
A list of the chosen data type - one entry for each sample
}
\description{
Get data is more generic method for retrieving data from the object
and returns a list, the number of entries in the list is number of
samples in the dataset and the list contain entries of the data type
and analysis stage you select.
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

getData(cds, 'matrixList','original')

}
