% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{getInteractions}
\alias{getInteractions}
\alias{getInteractions,comradesDataSet-method}
\title{getInteractions}
\usage{
getInteractions(cds, interactor)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{interactor}{The rna to show interactions with}
}
\description{
This method prints interactions of an RNA on the RNA of interest.
}
