% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{getInteractions}
\alias{getInteractions}
\alias{getInteractions,rnaCrosslinkDataSet-method}
\title{getInteractions}
\usage{
getInteractions(cds, interactors)
}
\arguments{
\item{cds}{a \code{rnaCrosslinkDataSet} object}

\item{interactors}{A vector containing the names of RNAs to show interactions with}
}
\value{
A table showing the read coverage of the specified interacting RNAs
}
\description{
This method returns a table of interactions of an RNA (interactor) on the RNA of interest.
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
getInteractions(cds, c("transcript1","transcript2"))

}
