% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonHelpersAndMethods.R
\docType{methods}
\name{getMatrices}
\alias{getMatrices}
\title{getMatrices}
\usage{
getMatrices(InputList, rna, size)
}
\arguments{
\item{InputList}{the original InputList created with readInputFiles or subsetInputList}

\item{rna}{the RNA of interest that you want to subset}

\item{size}{The size of the RNA}
}
\value{
A list of matrices
}
\description{
Make a matrix of contact interactions
}
\details{
Function used to create a list of matrices for plotting with
plotMatrixList or plotMatrixListFull, the output list will be same as the
input except for an extra list layer for the specific RNA
}
