% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{getReverseInteractions}
\alias{getReverseInteractions}
\alias{getReverseInteractions,rnaCrosslinkDataSet-method}
\title{getReverseInteractions}
\usage{
getReverseInteractions(cds, interactor)
}
\arguments{
\item{cds}{a \code{rnaCrosslinkDataSet} object}

\item{interactor}{The rna to show interactions with}
}
\value{
A long format table shoing the read coverage of chosen RNA
}
\description{
This method prints interactions of the RNA of interest on another RNA
transcript.
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
getReverseInteractions(cds, 'transcript2')

}
