% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\name{group}
\alias{group}
\alias{group,rnaCrosslinkDataSet-method}
\title{group}
\usage{
group(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A list - The indices of the sample in the control and sample groups
}
\description{
Extract the indeces for each group for the instance
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

group(cds)
}
