% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{hybFiles}
\alias{hybFiles}
\alias{hybFiles,comradesDataSet-method}
\title{hybFiles}
\usage{
hybFiles(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\value{
A list of tables in the original input format, one entry for each sample
}
\description{
Extract the data in original format
}
\examples{
cds = makeExampleComradesDataSet()

hybFiles(cds)
}
