% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{matrixList}
\alias{matrixList}
\alias{matrixList,comradesDataSet-method}
\title{matrixList}
\usage{
matrixList(x)
}
\arguments{
\item{x}{A comradesDataSet object}
}
\value{
A list of contract matrices, one entry for each sample
}
\description{
Extract the contact matrices
}
\examples{
cds = makeExampleComradesDataSet()

matrixList(cds)
}
