% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{matrixList<-}
\alias{matrixList<-}
\alias{matrixList<-,rnaCrosslinkDataSet-method}
\title{matrixList}
\usage{
matrixList(x) <- value
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}

\item{value}{A replacement}
}
\value{
No return - Sets a new matrixList slot
}
\description{
Set new matrixList slot
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

newMatrixList <- matrixList(cds)
matrixList(cds) <- newMatrixList
}
