% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonStatsAndPlots.R
\docType{methods}
\name{plotMatricesAverage}
\alias{plotMatricesAverage}
\alias{plotMatricesAverage,comradesDataSet-method}
\title{plotMatricesAverage}
\usage{
plotMatricesAverage(cds, type, directory, a, b, c, d, h)
}
\arguments{
\item{cds}{A comradesDataSet object}

\item{type}{The analysis stage to plot}

\item{directory}{An output directory for the heatmap (use 0 for no output)}

\item{a}{To make a subsetted plot (left value on x)}

\item{b}{To make a subsetted plot (right value on x)}

\item{c}{To make a subsetted plot (left value on y)}

\item{d}{To make a subsetted plot (right value on y)}

\item{h}{Height of image (inches) ( only useful if plotting)}
}
\description{
Plots a contact map to file of all samples for a stage in the analysis
}
