% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldComradesMethodsAndHelpers.R
\docType{methods}
\name{plotStructure}
\alias{plotStructure}
\alias{plotStructure,comradesDataSet-method}
\title{plotStructure}
\usage{
plotStructure(foldedCds, rnaRefs, s = "s1", n = 1)
}
\arguments{
\item{foldedCds}{\code{comradesDataSet} after running foldComrades}

\item{rnaRefs}{A fasta of the transcript (made with seqinr::read.fasta)}

\item{s}{sample of structure}

\item{n}{number of structure}
}
\description{
This method plots a structures chosen from the
plotEnsemblePCA method
}
