% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonStatsAndPlots.R
\docType{methods}
\name{readNumbers}
\alias{readNumbers}
\alias{readNumbers,rnaCrosslinkDataSet-method}
\title{readNumbers}
\usage{
readNumbers(knowClusteredCds, rna)
}
\arguments{
\item{knowClusteredCds}{A rnaCrosslinkDataSet object after clustering has been performed}

\item{rna}{The RNA ID of interest - use rna(cdsObject).}
}
\value{
A data.frame shoing the number of reads in clusters for each sample
}
\description{
This method prints a table showing the number of duplexes in 
the clusters in each step of the analysis
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

clusteredCds = clusterrnaCrosslink(cds,
                cores = 1,
                stepCount = 1,
                clusterCutoff = 1)
readNumbers(clusteredCds)
}
