% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{rnaSize}
\alias{rnaSize}
\alias{rnaSize,rnaCrosslinkDataSet-method}
\title{rnaSize}
\usage{
rnaSize(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A numeric - the size of the RNA (nucleotides)
}
\description{
Extract the size of the RNA for the instance
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

rnaSize(cds)
}
