% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,rnaCrosslinkDataSet-method}
\title{sampleNames}
\usage{
sampleNames(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A character vector - the sample names
}
\description{
Extract the sample names for the instance
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

sampleNames(cds)
}
