% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericMethods.R
\docType{methods}
\name{sampleTable}
\alias{sampleTable}
\alias{sampleTable,rnaCrosslinkDataSet-method}
\title{sampleTable}
\usage{
sampleTable(x)
}
\arguments{
\item{x}{A rnaCrosslinkDataSet object}
}
\value{
A data frame - The orginal meta-data table
}
\description{
Extract the sample table for the instance
}
\examples{
cds = makeExamplernaCrosslinkDataSet()

sampleTable(cds)
}
