% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comradesDataSetMethodsAndHelpers.R
\docType{methods}
\name{topInteracters}
\alias{topInteracters}
\alias{topInteracters,comradesDataSet-method}
\title{topInteracters}
\usage{
topInteracters(cds, ntop = 10)
}
\arguments{
\item{cds}{a \code{comradesDataSet} object}

\item{ntop}{the number of entries to display}
}
\description{
This method prints the top transcripts that have the most duplexes
assigned
}
