% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{topInteracters}
\alias{topInteracters}
\alias{topInteracters,rnaCrosslinkDataSet-method}
\title{topInteracters}
\usage{
topInteracters(cds, ntop = 10)
}
\arguments{
\item{cds}{a \code{rnaCrosslinkDataSet} object}

\item{ntop}{the number of entries to display}
}
\value{
A table, the number of counts per sample per interacting transcript
}
\description{
This method prints the top transcripts that have the most duplexes
assigned that interact with the transcript of interest
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
topInteracters(cds)

}
