% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaCrosslinkDataSetMethodsAndHelpers.R
\docType{methods}
\name{topInteractions}
\alias{topInteractions}
\alias{topInteractions,rnaCrosslinkDataSet-method}
\title{topInteractions}
\usage{
topInteractions(cds, ntop = 10)
}
\arguments{
\item{cds}{a \code{rnaCrosslinkDataSet} object}

\item{ntop}{the number of entries to display}
}
\value{
A table, the number of counts per sample per interaction
}
\description{
This method prints the top transcript interactions that have the most duplexes
assigned
}
\examples{
cds = makeExamplernaCrosslinkDataSet()
topInteractions(cds)

}
